Try a different method for general-key-dispatch
This is with regards to #18.

For some reason defining the fallback command in a map to an arbitrary
key and using set-transient-map and unread-command-events does not work
consistently depending on what key the fallback command is bound to. I'm
not sure if this has something to do with how
overriding-terminal-local-map works, but while it works consistently for
me after changing the initial key, it seems this is not the case for
others. It is possible that the original method could still work if I
figure out the reason for this, but this new method is probably cleaner
anyway.