Improve repeat handling of general-key-dispatch
Repeating should now work both for mapped commands and the fallback
command. Previously, repeating would only work would only work for the
fallback command if the created function was bound to a single key. Now
evil-repeat-info is modified correctly regardless.

The lambda keyword keyword is removed, so that the repeat property can
be set for created functions. An alternative would be to modify
evil-repeat-info completely from within the function generated by
general-key-dispatch. One problem with this is that it seems that
changing evil-repeat-info for a mapped command does not work. It must be
changed later (e.g. by advising evil-repeat-stop).

Also, commands that should not be repeated (e.g. motions in normal
state) are no longer be repeated.

Related to #19.