Set evil-this-operator in general-dispatch-...
evil-define-operator sets evil-this-operator to this-command. When
calling an operator interactively (e.g. evil-change), this-command will
not be correct, so commands that check evil-this-operator will not work
correctly. Temporarily setting this-command to the command being called
fixes this without breaking repeating (the general--fix-repeat function
won't be called if this-command changes). This allows evil-forward-word
to work correctly when evil-want-change-word-to-end is non-nil and for
evil-surround to automatically work with a general-dispatch-...
command (fixes #20).