Fix more broken repeat cases for dispatch/simulate
Addresses #19.

The previous implementations had a lot of problems and only handled the
most basic cases. Now general-key-dispatch commands that contain
general-simulate commands should work correctly. Counts of arbitrary
length in the middle of a general-key-dispatch command should work
correctly. Duplicate keys are not recorded regardless of the length of
the key sequences bound in the dispatch command. All tests now pass, and
there are no longer any significant (see the new NOTE) edge cases that I
am aware of, but some tests still need to be added. Only the end of
evil-repeat-info is now modified, so commands should work when they are
not first in a repeat chain, but there are currently no tests for this.

If a command name is given in a simulate command, its repeat properties
are checked to determine whether to abort, but if a name is not given,
the user must manually set the repeat property of the general-simulate
command if they do not want it to be repeated.

general-key-dispatch no longer immediately reads in a char (it no longer
has the "c" interactive code), so it is now possible to implement #42.